requirements.txt
youtube_transcript_api
python-dotenv